codingChallenge fun and messing with console colors
