+++ src/rpg_test/Action.java
@@ -1,0 +1,4 @@
+package rpg_test;
+
+public class Action {
+}
