+++ src/rpg_test/Companions.java
@@ -1,0 +1,4 @@
+package rpg_test;
+
+public class Companions {
+}
