+++ src/rpg_test/Stats.java
@@ -1,0 +1,101 @@
+package rpg_test;
+
+import tools.ConsoleColors;
+
+public class Stats {
+
+    public static void pathOfTheNomad() {
+        Toons.charclass = "Nomad";
+        Toons.hp = 20;
+        Toons.maxHp = 2000;
+
+        Toons.armor = 20;
+        Toons.maxArmor = 10000;
+
+        Toons.meleeDmg = 4;
+        Toons.maxMeleeDmg =
+
+        Toons.speed = 12;
+        Toons.maxSpeed = 25;
+
+        Toons.mana = 15;
+        Toons.maxMana = 47;
+
+        Toons.luck = 1;
+
+
+        Toons.xp = 0;
+        Toons.level = 1;
+    }
+
+    public static void pathOfTheAlchemist() {
+        Toons.charclass = "Alchemist";
+        Toons.hp = 20;
+        Toons.maxHp = 2000;
+
+        Toons.armor = 20;
+        Toons.maxArmor = 700;
+
+        Toons.meleeDmg = 4;
+        Toons.maxMeleeDmg = 34;
+
+        Toons.speed = 12;
+        Toons.maxSpeed = 25;
+
+        Toons.mana = 15;
+        Toons.maxMana = 47;
+
+        Toons.luck = 1;
+
+        Toons.xp = 0;
+        Toons.level = 1;
+    }
+
+    public static void pathOfTheBerserker() {
+        Toons.charclass = ConsoleColors.RED_BOLD + "Berserker";
+        Toons.hp = 20;
+        Toons.maxHp = 2000;
+
+        Toons.armor = 20;
+        Toons.maxArmor = 1000;
+
+        Toons.meleeDmg = 4;
+        Toons.maxMeleeDmg = 42;
+
+        Toons.speed = 12;
+        Toons.maxSpeed = 25;
+
+        Toons.mana = 15;
+        Toons.maxMana = 47;
+
+        Toons.luck = 1;
+
+        Toons.xp = 0;
+        Toons.level = 1;
+    }
+
+    private static void pathofTheArcher() {
+        Toons.charclass = "1";
+        Toons.hp = 20;
+        Toons.maxHp = 2000;
+
+        Toons.armor = 20;
+        Toons.maxArmor = 800;
+
+        Toons.meleeDmg = 4;
+        Toons.maxMeleeDmg =
+
+        Toons.speed = 12;
+        Toons.maxSpeed = 25;
+
+        Toons.mana = 15;
+        Toons.maxMana = 47;
+
+        Toons.luck = 1;
+
+        Toons.xp = 0;
+        Toons.level = 1;
+    }
+
+
+}
