+++ src/test/Main.java
@@ -1,7 +1,22 @@
 package test;
 
 
+import java.util.Scanner;
+
+
 public class Main {
+
+    public static String broDar() {
+
+//        Scanner scanner = new Scanner();
+//        scanner
+
+
+        return "bro ";
+    }
+
+
+
     public static void main(String[] args) {
