+++ src/test/Patterns.java
@@ -3,5 +3,5 @@ package test;
 public class Patterns {
 
 //    print the right amout (n1) of 'o'
-    // 
+    //
 }
