+++ src/tools/ConsoleColors.java
@@ -1,3 +1,5 @@
+package tools;
+
 public class ConsoleColors {
