+++ src/tools/dice.java
@@ -1,3 +1,5 @@
+package tools;
+
 import java.util.*;
 import java.util.Random;
